pyproject.toml
[project]
name = "intellisense-ai-smart-watch-web"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "channels (>=4.2.2,<5.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "openai (>=1.77.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "websockets (>=15.0.1,<16.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
